Core is implicitly allowed to scaffold files
Mention in the documentation that Drupal core is implicitly allowed to scaffold files. (drupal/legacy-scaffold-assets is too, but imho we should not be totally complete in the docs).

One allowed package example uses `drupal/core` as allowed-package, but that is not a realistic example. Therefore I propose to change it. Not having a realistic example I chose `example/assets`.